refactor: CR fix and not remove subscriptions from navigation component
