+++ projects/storefrontlib/src/lib/cms-lib/footer-navigation/footer-navigation.component.html
@@ -19,9 +19,6 @@
     </div>
   </div>
 </nav>
-<div
-  class="notice"
-  *ngIf="(component.data$ | async) as data"
->
+<div class="notice" *ngIf="(component.data$ | async) as data">
   {{ data.notice }}
 </div>
