+++ projects/storefrontlib/src/lib/cms-lib/footer-navigation/footer-navigation.component.spec.ts
@@ -10,7 +10,7 @@ import { NavigationService } from '../navigation/navigation.service';
 import { NavigationComponent } from '..';
 import { FooterNavigationComponent } from './footer-navigation.component';
 import { of } from 'rxjs';
-import { CmsComponentData } from '../../cms';
+import { CmsComponentData } from '../../cms/components/cms-component-data';
 
 @Component({
   selector: 'cx-navigation-ui',
