+++ projects/storefrontlib/src/lib/cms-lib/navigation/navigation-node.model.ts
@@ -1,0 +1,6 @@
+export interface NavigationNode {
+  title?: string;
+  url?: string;
+  target?: string | boolean;
+  children?: Array<NavigationNode>;
+}
